Backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
pytest==7.2.2
rest_framework_simplejwt==0.0.1
